requirements.txt
flask
Flask-SQLAlchemy
flask_smorest
marshmallow
passlib
flask_jwt_extended
python-dotenv
validators
shortuuid